requirements.txt
Django>=5.1, <6.0
djangorestframework>=3.15.2,<4.0
psycopg2>=2.9.5,<3.0
drf-spectacular>=0.27.0, <0.28.0
drf-schema-adapter>=3.0.6
